Removed old rcfile parser
